+++ packages/itwin/tree-widget/src/test/trees/models-tree/Utils.ts
@@ -21,6 +21,7 @@ import type { IModelConnection } from "@itwin/core-frontend";
 import type {
   ClassGroupingNodeKey,
   GroupingHierarchyNode,
+  GroupingNodeKey,
   HierarchyFilteringPath,
   HierarchyNodeKey,
   HierarchyProvider,
@@ -33,21 +34,25 @@ interface CreateModelsTreeProviderProps {
   imodel: IModelConnection;
   filteredNodePaths?: HierarchyFilteringPath[];
   hierarchyConfig?: Partial<ModelsTreeHierarchyConfiguration>;
+  idsCache?: ModelsTreeIdsCache;
+  imodelAccess?: ReturnType<typeof createIModelAccess>;
 }
 
 export function createModelsTreeProvider({
   imodel,
   filteredNodePaths,
   hierarchyConfig,
+  imodelAccess,
+  idsCache,
 }: CreateModelsTreeProviderProps): HierarchyProvider & { dispose: () => void; [Symbol.dispose]: () => void } {
   const config = { ...defaultHierarchyConfiguration, hideRootSubject: true, ...hierarchyConfig };
-  const imodelAccess = createIModelAccess(imodel);
-  const idsCache = new ModelsTreeIdsCache(imodelAccess, config);
+  const createdImodelAccess = imodelAccess ?? createIModelAccess(imodel);
+  const createdIdsCache = idsCache ?? new ModelsTreeIdsCache(createdImodelAccess, config);
   const provider = createIModelHierarchyProvider({
-    imodelAccess,
+    imodelAccess: createdImodelAccess,
     hierarchyDefinition: new ModelsTreeDefinition({
-      imodelAccess,
-      idsCache,
+      imodelAccess: createdImodelAccess,
+      idsCache: createdIdsCache,
       hierarchyConfig: config,
     }),
     ...(filteredNodePaths
@@ -56,7 +61,9 @@ export function createModelsTreeProvider({
   });
   const dispose = () => {
     provider[Symbol.dispose]();
-    idsCache[Symbol.dispose]();
+    if (!idsCache) {
+      createdIdsCache[Symbol.dispose]();
+    }
   };
   return {
     hierarchyChanged: provider.hierarchyChanged,
@@ -199,4 +206,25 @@ export function createClassGroupingHierarchyNode({
     parentKeys: parentKeys ?? [],
     extendedData: props,
   };
+}
+
+export function getNodeParentKeys(
+  keys: (
+    | GroupingNodeKey
+    | {
+        className: string;
+        id: string;
+      }
+  )[],
+): HierarchyNodeKey[] {
+  return keys.map((key) => {
+    if ("type" in key) {
+      return key;
+    }
+
+    return {
+      type: "instances",
+      instanceKeys: [key],
+    };
+  });
 }