+++ packages/itwin/tree-widget/src/tree-widget-react/components/trees/classifications-tree/ClassificationsTree.tsx
@@ -4,19 +4,16 @@
  *--------------------------------------------------------------------------------------------*/
 
 import type { VisibilityTreeProps } from "../common/components/VisibilityTree.js";
-import { useCallback } from "react";
-import { RenameAction } from "@itwin/presentation-hierarchies-react";
 import { VisibilityTree } from "../common/components/VisibilityTree.js";
 import { VisibilityTreeRenderer } from "../common/components/VisibilityTreeRenderer.js";
 import { useClassificationsTree } from "./UseClassificationsTree.js";
 
-import type { PresentationHierarchyNode } from "@itwin/presentation-hierarchies-react";
 import type { VisibilityTreeRendererProps } from "../common/components/VisibilityTreeRenderer.js";
 import type { UseClassificationsTreeProps } from "./UseClassificationsTree.js";
 
 /** @alpha */
 export type ClassificationsTreeProps = Pick<VisibilityTreeProps, "imodel" | "selectionStorage" | "selectionMode" | "emptyTreeContent"> &
-  Pick<VisibilityTreeRendererProps, "getActions" | "getDecorations" | "getEditingProps"> &
+  Pick<VisibilityTreeRendererProps, "getInlineActions" | "getMenuActions" | "getDecorations" | "getEditingProps"> &
   UseClassificationsTreeProps & {
     hierarchyLevelConfig?: {
       sizeLimit?: number;
@@ -33,7 +30,8 @@ export function ClassificationsTree({
   selectionMode,
   emptyTreeContent,
   getDecorations,
-  getActions,
+  getInlineActions,
+  getMenuActions,
   getEditingProps,
   filter,
 }: ClassificationsTreeProps) {
@@ -44,12 +42,12 @@ export function ClassificationsTree({
     filter,
   });
 
-  const nodeActions = useCallback(
-    (node: PresentationHierarchyNode) => {
-      return [<RenameAction key="RenameAction" />, ...(getActions ? getActions(node) : [])];
-    },
-    [getActions],
-  );
+  // const nodeActions = useCallback(
+  //   (node: PresentationHierarchyNode) => {
+  //     return [<RenameAction key="RenameAction" />, ...(getActions ? getActions(node) : [])];
+  //   },
+  //   [getActions],
+  // );
 
   return (
     <VisibilityTree
@@ -63,7 +61,8 @@ export function ClassificationsTree({
           {...treeProps}
           {...rendererProps}
           getEditingProps={getEditingProps}
-          getActions={nodeActions}
+          getInlineActions={getInlineActions}
+          getMenuActions={getMenuActions}
           getDecorations={getDecorations ?? rendererProps.getDecorations}
         />
       )}