+++ packages/itwin/tree-widget/src/tree-widget-react/components/trees/classifications-tree/ClassificationsTreeComponent.tsx
@@ -18,7 +18,16 @@ import type { IModelConnection, ScreenViewport } from "@itwin/core-frontend";
 interface ClassificationsTreeComponentProps
   extends Pick<
     ClassificationsTreeProps,
-    "selectionStorage" | "hierarchyLevelConfig" | "selectionMode" | "filter" | "emptyTreeContent" | "getActions" | "getDecorations" | "hierarchyConfig" | "getEditingProps"
+    | "selectionStorage"
+    | "hierarchyLevelConfig"
+    | "selectionMode"
+    | "filter"
+    | "emptyTreeContent"
+    | "getInlineActions"
+    | "getMenuActions"
+    | "getDecorations"
+    | "hierarchyConfig"
+    | "getEditingProps"
   > {
   onPerformanceMeasured?: (featureId: string, duration: number) => void;
   onFeatureUsed?: (feature: string) => void;