+++ packages/itwin/tree-widget/src/tree-widget-react/components/trees/common/components/TreeNodeVisibilityButton.tsx
@@ -4,15 +4,16 @@
  *--------------------------------------------------------------------------------------------*/
 
 import "./TreeNodeVisibilityButton.css";
-import { memo } from "react";
+import { createContext, memo, useContext, useMemo } from "react";
 import visibilityHideSvg from "@stratakit/icons/visibility-hide.svg";
 import visibilityPartialSvg from "@stratakit/icons/visibility-partial.svg";
 import visibilityShowSvg from "@stratakit/icons/visibility-show.svg";
 import { Tree } from "@stratakit/structures";
+import { TreeWidget } from "../../../../TreeWidget.js";
 import { createTooltip } from "../internal/Tooltip.js";
 
+import type { PropsWithChildren } from "react";
 import type { PresentationHierarchyNode } from "@itwin/presentation-hierarchies-react";
-
 /**
  * Data structure that describes tree node checkbox state.
  * @beta
@@ -25,19 +26,30 @@ export interface TreeItemVisibilityButtonState {
 
 /** @beta */
 export interface TreeItemVisibilityButtonProps {
-  /** Callback that should be invoked when checkbox is clicked. */
-  onVisibilityButtonClick: (node: PresentationHierarchyNode, state: TreeItemVisibilityButtonState["state"]) => void;
-  /** Callback that should be used to determine current checkbox state. */
-  getVisibilityButtonState: (node: PresentationHierarchyNode) => TreeItemVisibilityButtonState;
+  /**
+   * Indicates that space for this action button should be reserved, even when the action is not available.
+   * For nodes that don't support visibility, `<VisibilityAction reserveSpace />` renders:
+   *
+   * - Blank space when the action is used as an inline action. It's recommended to set this prop to keep all action buttons of the same kind vertically aligned.
+   * - Disabled menu item when the action is used as a menu action.
+   */
+  reserveSpace?: true;
 }
 
-/** @internal */
-export const VisibilityAction = memo(function VisibilityAction({
-  getVisibilityButtonState,
-  onVisibilityButtonClick,
-  node,
-}: TreeItemVisibilityButtonProps & { node: PresentationHierarchyNode }) {
-  const state = getVisibilityButtonState(node);
+/**
+ * React component that renders a visibility action for a tree item.
+ * Should be used with `VisibilityTreeRenderer`.
+ * @beta
+ */
+export const VisibilityAction = memo(function VisibilityAction({ node, reserveSpace }: TreeItemVisibilityButtonProps & { node: PresentationHierarchyNode }) {
+  const context = useVisibilityContext();
+  const state = context?.getVisibilityButtonState(node);
+
+  if (!context || !state || state.isDisabled) {
+    return reserveSpace ? (
+      <Tree.ItemAction label={TreeWidget.translate(`visibilityTooltips.status.disabled`)} visible={false} icon={visibilityShowSvg} />
+    ) : undefined;
+  }
 
   const getIcon = () => {
     switch (state.state) {
@@ -53,9 +65,41 @@ export const VisibilityAction = memo(function VisibilityAction({
   return (
     <Tree.ItemAction
       label={state.tooltip ?? createTooltip(state.state)}
-      onClick={() => onVisibilityButtonClick(node, state.state)}
-      visible={state.isDisabled ? false : state.state !== "visible" ? true : undefined}
+      onClick={() => context.onVisibilityButtonClick(node, state.state)}
+      visible={state.state !== "visible" ? true : undefined}
       icon={getIcon()}
     />
   );
 });
+
+/** @beta */
+export interface VisibilityContext {
+  /** Callback that should be invoked when checkbox is clicked. */
+  onVisibilityButtonClick: (node: PresentationHierarchyNode, state: TreeItemVisibilityButtonState["state"]) => void;
+  /** Callback that should be used to determine current checkbox state. */
+  getVisibilityButtonState: (node: PresentationHierarchyNode) => TreeItemVisibilityButtonState;
+}
+
+const visibilityContext = createContext<VisibilityContext | undefined>(undefined);
+
+/** @internal */
+export const useVisibilityContext = () => {
+  return useContext(visibilityContext);
+};
+
+/** @internal */
+export function VisibilityContextProvider({ onVisibilityButtonClick, getVisibilityButtonState, children }: PropsWithChildren<VisibilityContext>) {
+  return (
+    <visibilityContext.Provider
+      value={useMemo(
+        () => ({
+          onVisibilityButtonClick,
+          getVisibilityButtonState,
+        }),
+        [getVisibilityButtonState, onVisibilityButtonClick],
+      )}
+    >
+      {children}
+    </visibilityContext.Provider>
+  );
+}