+++ packages/itwin/tree-widget/src/tree-widget-react/components/trees/common/components/VisibilityTree.tsx
@@ -11,9 +11,9 @@ import { Tree } from "./Tree.js";
 import type { TreeProps } from "./Tree.js";
 import type { FunctionProps } from "../Utils.js";
 import type { ReactNode } from "react";
-import type { VisibilityTreeRendererProps } from "./VisibilityTreeRenderer.js";
 import type { ECClassHierarchyInspector } from "@itwin/presentation-shared";
 import type { HierarchyVisibilityHandler } from "../UseHierarchyVisibility.js";
+import type { VisibilityContext } from "./TreeNodeVisibilityButton.js";
 
 /** @beta */
 export type VisibilityTreeProps = Omit<TreeProps, "treeRenderer" | "imodelAccess"> & {
@@ -21,7 +21,7 @@ export type VisibilityTreeProps = Omit<TreeProps, "treeRenderer" | "imodelAccess
   visibilityHandlerFactory: (props: { imodelAccess: ECClassHierarchyInspector }) => HierarchyVisibilityHandler;
   /** Tree renderer that should be used to render tree data. */
   treeRenderer: (
-    treeProps: FunctionProps<TreeProps["treeRenderer"]> & Pick<VisibilityTreeRendererProps, "getVisibilityButtonState" | "onVisibilityButtonClick">,
+    treeProps: FunctionProps<TreeProps["treeRenderer"]> & Pick<VisibilityContext, "getVisibilityButtonState" | "onVisibilityButtonClick">,
   ) => ReactNode;
 };