+++ packages/itwin/tree-widget/src/tree-widget-react/components/trees/common/components/VisibilityTreeRenderer.tsx
@@ -6,15 +6,15 @@
 import { useCallback } from "react";
 import { FilterAction } from "@itwin/presentation-hierarchies-react";
 import { BaseTreeRenderer } from "./BaseTreeRenderer.js";
-import { VisibilityAction } from "./TreeNodeVisibilityButton.js";
+import { VisibilityAction, VisibilityContextProvider } from "./TreeNodeVisibilityButton.js";
 import { useVisibilityButtonHandler } from "./UseVisibilityButtonHandler.js";
 
 import type { PresentationHierarchyNode } from "@itwin/presentation-hierarchies-react";
 import type { BaseTreeRendererProps } from "./BaseTreeRenderer.js";
-import type { TreeItemVisibilityButtonProps } from "./TreeNodeVisibilityButton.js";
+import type { VisibilityContext } from "./TreeNodeVisibilityButton.js";
 
 /** @beta */
-export type VisibilityTreeRendererProps = BaseTreeRendererProps & TreeItemVisibilityButtonProps;
+export type VisibilityTreeRendererProps = BaseTreeRendererProps & VisibilityContext;
 
 /**
  * Tree renderer that renders tree nodes with eye checkboxes for controlling visibility of instances represented by tree nodes.
@@ -23,7 +23,7 @@ export type VisibilityTreeRendererProps = BaseTreeRendererProps & TreeItemVisibi
 export function VisibilityTreeRenderer({
   getVisibilityButtonState,
   onVisibilityButtonClick: onClick,
-  getActions,
+  getInlineActions,
   onFilterClick,
   getHierarchyLevelDetails,
   ...props
@@ -32,19 +32,19 @@ export function VisibilityTreeRenderer({
 
   const nodeActions = useCallback(
     (node: PresentationHierarchyNode) => {
-      return [
-        <VisibilityAction
-          key={"Visibility"}
-          node={node}
-          onVisibilityButtonClick={onVisibilityButtonClick}
-          getVisibilityButtonState={getVisibilityButtonState}
-        />,
-        <FilterAction key={"Filter"} node={node} onFilter={onFilterClick} getHierarchyLevelDetails={getHierarchyLevelDetails} />,
-        ...(getActions ? getActions(node) : []),
-      ];
+      return getInlineActions
+        ? getInlineActions(node)
+        : [
+            <VisibilityAction key={"Visibility"} node={node} reserveSpace />,
+            <FilterAction key={"Filter"} node={node} onFilter={onFilterClick} getHierarchyLevelDetails={getHierarchyLevelDetails} reserveSpace />,
+          ];
     },
-    [onVisibilityButtonClick, getVisibilityButtonState, onFilterClick, getHierarchyLevelDetails, getActions],
+    [onFilterClick, getHierarchyLevelDetails, getInlineActions],
   );
 
-  return <BaseTreeRenderer {...props} onFilterClick={onFilterClick} getHierarchyLevelDetails={getHierarchyLevelDetails} getActions={nodeActions} />;
+  return (
+    <VisibilityContextProvider onVisibilityButtonClick={onVisibilityButtonClick} getVisibilityButtonState={getVisibilityButtonState}>
+      <BaseTreeRenderer {...props} onFilterClick={onFilterClick} getHierarchyLevelDetails={getHierarchyLevelDetails} getInlineActions={nodeActions} />
+    </VisibilityContextProvider>
+  );
 }