+++ packages/itwin/tree-widget/src/tree-widget-react/components/trees/models-tree/ModelsTree.tsx
@@ -10,9 +10,10 @@ import { useModelsTree } from "./UseModelsTree.js";
 import type { VisibilityTreeRendererProps } from "../common/components/VisibilityTreeRenderer.js";
 import type { UseModelsTreeProps } from "./UseModelsTree.js";
 import type { VisibilityTreeProps } from "../common/components/VisibilityTree.js";
+
 /** @beta */
 export type ModelsTreeProps = Pick<VisibilityTreeProps, "imodel" | "selectionStorage" | "selectionMode" | "emptyTreeContent"> &
-  Pick<VisibilityTreeRendererProps, "getActions" | "getDecorations"> &
+  Pick<VisibilityTreeRendererProps, "getInlineActions" | "getMenuActions" | "getDecorations"> &
   UseModelsTreeProps & {
     hierarchyLevelConfig?: {
       sizeLimit?: number;
@@ -33,8 +34,10 @@ export function ModelsTree({
   getFilteredPaths,
   getDecorations,
   onModelsFiltered,
-  getActions,
+  getInlineActions,
+  getMenuActions,
   emptyTreeContent,
+  getSubTreePaths,
 }: ModelsTreeProps) {
   const { modelsTreeProps, rendererProps } = useModelsTree({
     activeView,
@@ -45,6 +48,7 @@ export function ModelsTree({
     onModelsFiltered,
     selectionPredicate,
     emptyTreeContent,
+    getSubTreePaths,
   });
 
   return (
@@ -55,7 +59,13 @@ export function ModelsTree({
       hierarchyLevelSizeLimit={hierarchyLevelConfig?.sizeLimit}
       selectionMode={selectionMode}
       treeRenderer={(treeProps) => (
-        <VisibilityTreeRenderer {...treeProps} {...rendererProps} getActions={getActions} getDecorations={getDecorations ?? rendererProps.getDecorations} />
+        <VisibilityTreeRenderer
+          {...treeProps}
+          {...rendererProps}
+          getInlineActions={getInlineActions}
+          getMenuActions={getMenuActions}
+          getDecorations={getDecorations ?? rendererProps.getDecorations}
+        />
       )}
     />
   );