+++ packages/itwin/tree-widget/src/tree-widget-react/components/trees/models-tree/ModelsTreeComponent.tsx
@@ -39,8 +39,10 @@ interface ModelsTreeComponentProps
     | "getFilteredPaths"
     | "filter"
     | "emptyTreeContent"
-    | "getActions"
+    | "getInlineActions"
+    | "getMenuActions"
     | "getDecorations"
+    | "getSubTreePaths"
   > {
   /**
    * Renderers of header buttons. Defaults to: