+++ packages/itwin/geo-tools/CHANGELOG.md
@@ -1,8 +1,16 @@
 # Change Log - @itwin/geo-tools-react
 
-This log was last generated on Fri, 20 Jun 2025 17:05:06 GMT and should not be manually modified.
+<!-- This log was last generated on Mon, 07 Jul 2025 19:55:02 GMT and should not be manually modified. -->
 
 <!-- Start content -->
+
+## 6.0.1
+
+Mon, 07 Jul 2025 19:55:02 GMT
+
+### Patches
+
+- Removed unused `redux` peer dependency. ([#1368](https://github.com/iTwin/viewer-components-react/pull/1368))
 
 ## 6.0.0